feat(selector4nix): add `expire_at` field to `NarInfo`
