+++ src/application/nar_info/actor/runner.rs
@@ -1,8 +1,10 @@
 use std::sync::Arc;
+use std::time::{Duration, SystemTime};
 
 use selector4nix_actor::actor::{Actor, ActorPre, ActorPreBuilder, Context, EmptyInternal};
 use tokio::sync::oneshot::Sender as OneshotSender;
 
+use crate::domain::common::expire_at::ExpireAt;
 use crate::domain::nar_info::model::{NarInfo, ProxyNarInfoData};
 use crate::domain::nar_info::{NarInfoService, ResolveNarInfoError, ResolveNarInfoEvent};
 
@@ -30,14 +32,20 @@ pub struct NarInfoActor {
     init: Option<NarInfo>,
     context: Context<NarInfoRequest, EmptyInternal>,
     nar_info_service: Arc<NarInfoService>,
+    nar_info_ttl: Duration,
 }
 
 impl NarInfoActor {
-    pub fn new(init: NarInfo, nar_info_service: Arc<NarInfoService>) -> ActorPre<Self> {
+    pub fn new(
+        init: NarInfo,
+        nar_info_service: Arc<NarInfoService>,
+        nar_info_ttl: Duration,
+    ) -> ActorPre<Self> {
         ActorPreBuilder::inject(|context| Self {
             init: Some(init),
             context,
             nar_info_service,
+            nar_info_ttl,
         })
     }
 
@@ -46,6 +54,8 @@ impl NarInfoActor {
         nar: NarInfo,
         reply_to: OneshotSender<ResolveNarInfoResponse>,
     ) -> NarInfo {
+        let nar = nar.check_expiry_and_update(SystemTime::now());
+
         if let Some(resolution) = nar.resolution() {
             let res = Ok(resolution.nar_info().cloned());
             let _ = reply_to.send(ResolveNarInfoResponse::new(res, Vec::new()));
@@ -56,7 +66,8 @@ impl NarInfoActor {
         match res {
             Ok(resolution) => {
                 let res = Ok(resolution.nar_info().cloned());
-                let nar = nar.on_resolved(resolution);
+                let expire_at = ExpireAt::since(SystemTime::now(), self.nar_info_ttl);
+                let nar = nar.on_resolved(resolution, expire_at);
                 let _ = reply_to.send(ResolveNarInfoResponse::new(res, events));
                 nar
             }
