+++ src/domain/nar_info/model/mod.rs
@@ -1,11 +1,13 @@
 mod nar_file_name;
 mod nar_info;
+mod nar_info_resolution;
 mod proxy_nar_info_data;
 mod store_path_hash;
 mod upstream_nar_info_data;
 
 pub use nar_file_name::{NarFileName, TryNewNarFileNameError};
-pub use nar_info::{NarInfo, NarInfoResolution, NarUrlRewriteOption};
+pub use nar_info::NarInfo;
+pub use nar_info_resolution::{NarInfoResolution, NarUrlRewriteOption};
 pub use proxy_nar_info_data::ProxyNarInfoData;
 pub use store_path_hash::{StorePathHash, TryNewStorePathHashError};
 pub use upstream_nar_info_data::{TryUpstreamNewNarInfoData, UpstreamNarInfoData};
